requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
xgboost==2.0.0
scikit-learn==1.3.0
scipy==1.11.2
joblib==1.3.1
openpyxl==3.1.2
gunicorn==21.2.0
